+++ src/nextjs.ts
@@ -1,4 +1,6 @@
 export { createTcmOAuthExchangeRoute } from './nextjs/createTcmOAuthExchangeRoute';
 export { createTcmLogoutRoute } from './nextjs/createTcmLogoutRoute';
+export { refreshTcmAccessToken, revokeTcmToken } from './server/exchange';
 export type { CreateTcmOAuthExchangeRouteOptions } from './server/types';
 export type { CreateTcmLogoutRouteOptions } from './server/types';
+export type { RefreshTcmAccessTokenOptions, RevokeTcmTokenOptions } from './server/types';